+++ pc_backup/models/__init__.py
@@ -1,1 +1,1 @@
-import backup_computer, res_users, backup_odoo
+import backup_computer, res_users, backup_odoo, backup_settings