+++ pc_backup/models/backup_odoo.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-
 import datetime
 
 from openerp import api, fields, models
@@ -14,12 +13,3 @@ class BackupOdoo(models.Model):
     database_name = fields.Char(string="Database Name")
     data = fields.Binary(string="Data")
     data_filename = fields.Char("Data Filename")
-    
-    @api.model
-    def backup_cleanup(self):
-        
-        backup_days_to_keep = 7
-        
-        for backup in self.env['backup.odoo'].sudo().search([]):
-            if backup.create_date < (datetime.datetime.now() - datetime.timedelta(days=backup_days_to_keep) ).strftime("%Y-%m-%d %H:%M:%S"):
-                backup.unlink()